+++ src/geometry/quad.hpp
@@ -15,7 +15,7 @@ namespace ar
     Color color;
 
     // Creates a quad. Center & normal are expected to be ordered: x, y, z
-    Quad(float* center, float* normal) : width(0.0f), height(0.0f), color()
+    Quad(double* center, double* normal) : width(0.0f), height(0.0f), color()
     {
       for (int i = 0; i < 3; i++)
       {
@@ -28,7 +28,7 @@ namespace ar
     }
 
     // Creates a quad. Center & normal are expected to be ordered: x, y, z
-    Quad(float* center, float* normal, float width, float height) : color()
+    Quad(double* center, double* normal, double width, double height) : color()
     {
       for (int i = 0; i < 3; i++)
       {
@@ -41,7 +41,7 @@ namespace ar
     }
 
     // Creates a quad. Center & normal are expected to be ordered: x, y, z
-    Quad(float* center, float* normal, float width, float height, Color color)
+    Quad(double* center, double* normal, double width, double height, Color color)
     {
       for (int i = 0; i < 3; i++)
       {
